⚡️ :: MviViewModel, Reducer 병합
